magstats/pyproject.toml
[tool.poetry]
name = "magstats"
version = "0.1.0"
description = "Get Magstats of objects from ZTF and ATLAS surveys"
authors = ["mmolina2018 <mmolina2018@inf.udec.cl>"]
readme = "README.md"
packages = [{include = "core"}, {include = "api"}, {include = "database"}] 

[tool.poetry.dependencies]
fastapi = "^0.100.0"
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
returns = "^0.21.0"
psycopg2-binary = "^2.9.6"
db-plugins = { version = "5.1.1", python = ">=3.11,<3.12"}
prometheus-fastapi-instrumentator = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-docker = "^2.0.0"
black = "^23.7.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"

[tool.poetry.group.production.dependencies]
uvicorn = {extras = ["standard"], version = "^0.23.1"}

[tool.poetry.group.profiling.dependencies]
memray = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
